Move timestamp-estimation logic into TimestampEstimator class, and only do it from start until stop
Readout is not yet changed to send TimeSync messages from conf until
scrap, and it resets the timestamp at run start (to whatever the
timestamp of the first frame in the file was). This makes it hard to
do timestamp estimation correctly if we're doing it from conf until
scrap, so for now we just only do timestamp estimation from start
until stop.

To make it easier to switch back to conf-until-scrap timestamp
estimation, move the timestamp estimation logic into its own class,
TimestampEstimator, which estimates the current timestamp from
creation until destruction. If/when we want to change to
conf-until-scrap, we just move the creation of TimestampEstimator into
do_conf, and the destruction into do_scrap.